signate-competition/requirements.txt
# Core dependencies
numpy==1.24.3
pandas==2.1.4
scipy==1.11.4

# Machine Learning
torch==2.1.2
tensorflow==2.15.0
stable-baselines3==2.2.1

# Reinforcement Learning
gym==0.26.2
gymnasium==0.29.1

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Logging and monitoring
tensorboard==2.15.1
wandb==0.16.2
tqdm==4.66.1

# Configuration
pyyaml==6.0.1
omegaconf==2.3.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Performance profiling
memory-profiler==0.61.0
line-profiler==4.1.2

# Utilities
joblib==1.3.2
click==8.1.7
python-dotenv==1.0.0